fix ambiguous method bug #84

+++ src/Flee.NetStandard20/InternalTypes/Miscellaneous.cs
@@ -238,7 +238,7 @@ namespace Flee.InternalTypes
             // Our score is the average of the scores of each parameter.  The lower the score, the better the match.
             int sum = ComputeSum(parameters, argTypes);
 
-            return sum / argTypes.Length;
+            return (float)sum / (float)argTypes.Length;
         }
 
         private static int ComputeSum(ParameterInfo[] parameters, Type[] argTypes)
